feat(GraphQLTaskUtl): rename GraphQLTaskUtil to TaskMergeDataUtil
Rename GraphQLTaskUtil to TaskMergeDataUtil.

Closes DCOS-20068
